[BE] fix: 병합 충돌 해결 (주사위값 redis 저장, PlayerDiceMessage 주사위값들 Long 으로 변경, 더블 체크)

+++ backend/src/main/java/online/ft51land/modooseoul/domain/player/dto/message/PlayerDiceMessage.java
@@ -5,11 +5,11 @@ import online.ft51land.modooseoul.domain.player.entity.Player;
 
 @Builder
 public record PlayerDiceMessage (
-		int first,
-		int second,
+		Long first,
+		Long second,
 		Boolean isDouble
 ) {
-	public static PlayerDiceMessage of(int one, int two, Player player) {
+	public static PlayerDiceMessage of(Long one, Long two, Player player) {
 		return PlayerDiceMessage.builder()
 				.first(one).second(two).isDouble(player.getIsDouble())
 				.build();
